requirements.txt
certifi==2020.4.5.1
joblib==0.15.1
numpy==1.18.5
pandas==1.0.4
python-dateutil==2.8.1
pytz==2020.1
scikit-surprise==1.1.0
scipy==1.4.1
six==1.15.0
wincertstore==0.2
